Adds functionality to stop input from user and full words list.
The eventhandler associated with reading key presses is now removed
and only added again after the suggestion is displayed so that the user
is required to at least consider the suggestion, and so data is not
overwritten by the suggestion. The full list of over 12,000 words is
now included as well.