Adds clicking on suggestions to move to board.
Users can now click on a suggested word to move it to the gameboard,
rather than requiring them to type it in manually. Note that the
clicking on both the blocks and block container is disabled after the
stats modal shows up due to its current implementation. This should be
fixed once this is replaced with a box shadow.